+++ webofneeds/won-utils/won-utils-goals/src/test/resources/won/utils/goals/blending/differentProperties.trig
